Angelica - 12- Search ignores case and searches from start of word
The search now ignores case of the user's search and of the group names. It also searches for the user's search from the beginning of the group name: 
so, if a person searches "A" then only  Aquatics, Aerobics,  etc would come up instead of BasetbAll, BasebAll, etc...

Also took in consideration if the user's search is longer  than a group name. 

Will fix my methods and implement other features in future commits. This commit is just a draft.